+++ Sources/DeckKit/_Deprecated/Deprecated.swift
@@ -30,3 +30,57 @@ public extension Color {
         colorScheme == .light ? .white : .black
     }
 }
+
+public extension DeckView {
+    
+    @available(*, deprecated, message: "This is no longer used.")
+    typealias ItemAction = (ItemType) -> Void
+    
+    @available(*, deprecated, message: "Use the swipeAction initializer instead.")
+    init(
+        _ items: Binding<[ItemType]>,
+        shuffleAnimation: DeckShuffleAnimation = .init(),
+        swipeLeftAction: ItemAction? = nil,
+        swipeRightAction: ItemAction? = nil,
+        swipeUpAction: ItemAction? = nil,
+        swipeDownAction: ItemAction? = nil,
+        itemView: @escaping ItemViewBuilder
+    ) {
+        self.init(
+            items,
+            shuffleAnimation: shuffleAnimation,
+            swipeAction: { edge, item in
+                switch edge {
+                case .leading: swipeLeftAction?(item)
+                case .trailing: swipeRightAction?(item)
+                case .top: swipeUpAction?(item)
+                case .bottom: swipeDownAction?(item)
+                }
+            },
+            itemView: itemView
+        )
+    }
+    
+    @available(*, deprecated, message: "Apply a configuration with the .deckViewConfiguration view modifier instead.")
+    init(
+        _ items: Binding<[ItemType]>,
+        config: DeckViewConfiguration,
+        shuffleAnimation: DeckShuffleAnimation = .init(),
+        swipeLeftAction: ItemAction? = nil,
+        swipeRightAction: ItemAction? = nil,
+        swipeUpAction: ItemAction? = nil,
+        swipeDownAction: ItemAction? = nil,
+        itemView: @escaping ItemViewBuilder
+    ) {
+        self.init(
+            items,
+            shuffleAnimation: shuffleAnimation,
+            swipeLeftAction: swipeLeftAction,
+            swipeRightAction: swipeRightAction,
+            swipeUpAction: swipeUpAction,
+            swipeDownAction: swipeDownAction,
+            itemView: itemView
+        )
+        self.initConfig = config
+    }
+}